move duplicated node creation visitors into FileTypes.java; make methods static so we don't need to instantiate throw away object.
