+++ Core/src/org/sleuthkit/autopsy/datamodel/FileTypesByMimeType.java
@@ -31,7 +31,6 @@ import java.util.Observer;
 import java.util.logging.Level;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.openide.nodes.AbstractNode;
 import org.openide.nodes.ChildFactory;
 import org.openide.nodes.Children;
 import org.openide.nodes.Node;
@@ -42,13 +41,6 @@ import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.ingest.IngestManager;
 import org.sleuthkit.datamodel.AbstractFile;
 import org.sleuthkit.datamodel.Content;
-import org.sleuthkit.datamodel.ContentVisitor;
-import org.sleuthkit.datamodel.DerivedFile;
-import org.sleuthkit.datamodel.Directory;
-import org.sleuthkit.datamodel.File;
-import org.sleuthkit.datamodel.LayoutFile;
-import org.sleuthkit.datamodel.LocalFile;
-import org.sleuthkit.datamodel.SlackFile;
 import org.sleuthkit.datamodel.SleuthkitCase;
 import org.sleuthkit.datamodel.TskCoreException;
 import org.sleuthkit.datamodel.TskData;
@@ -358,10 +350,9 @@ public final class FileTypesByMimeType extends Observable implements AutopsyVisi
          *                 results
          */
         private void updateDisplayName(String mimeType) {
-            final long count = new MediaSubTypeNodeChildren(mimeType).calculateItems(skCase, mimeType);
-            String[] mimeTypeParts = mimeType.split("/");
-            //joins up all remaining parts of the mimeType into one sub-type string
-            super.setDisplayName(StringUtils.join(ArrayUtils.subarray(mimeTypeParts, 1, mimeTypeParts.length), "/") + " (" + count + ")");
+            final long count = calculateItems(skCase, mimeType);
+            //only the part of the mimeType after the media type
+            super.setDisplayName(StringUtils.substringAfter(mimeType, "/") + " (" + count + ")");
         }
 
         /**
@@ -392,6 +383,48 @@ public final class FileTypesByMimeType extends Observable implements AutopsyVisi
     }
 
     /**
+     * Get children count without actually loading all nodes
+     *
+     * @return count(*) - the number of items that will be shown in this items
+     *         Directory Listing
+     */
+    static private long calculateItems(SleuthkitCase sleuthkitCase, String mime_type) {
+        try {
+            return sleuthkitCase.countFilesWhere(createQuery(mime_type));
+        } catch (TskCoreException ex) {
+            LOGGER.log(Level.SEVERE, "Error getting file search view count", ex); //NON-NLS
+            return 0;
+        }
+    }
+
+    /**
+     * Create the portion of the query following WHERE for a query of the
+     * database for each file which matches the complete MIME type represented
+     * by this node. Matches against the mime_type column in tsk_files.
+     *
+     * @param mimeType - the complete mimetype of the file mediatype/subtype
+     *
+     * @return query.toString - portion of SQL query which will follow a WHERE
+     *         clause.
+     */
+    static private String createQuery(String mimeType) {
+        StringBuilder query = new StringBuilder();
+        query.append("(dir_type = ").append(TskData.TSK_FS_NAME_TYPE_ENUM.REG.getValue()).append(")"); //NON-NLS
+        query.append(" AND (type IN (").append(TskData.TSK_DB_FILES_TYPE_ENUM.FS.ordinal()).append(",");  //NON-NLS
+        query.append(TskData.TSK_DB_FILES_TYPE_ENUM.CARVED.ordinal()).append(",");
+        query.append(TskData.TSK_DB_FILES_TYPE_ENUM.DERIVED.ordinal()).append(",");
+        if (!UserPreferences.hideSlackFilesInViewsTree()) {
+            query.append(TskData.TSK_DB_FILES_TYPE_ENUM.SLACK.ordinal()).append(",");
+        }
+        query.append(TskData.TSK_DB_FILES_TYPE_ENUM.LOCAL.ordinal()).append("))");
+        if (UserPreferences.hideKnownFilesInViewsTree()) {
+            query.append(" AND (known IS NULL OR known != ").append(TskData.FileKnown.KNOWN.getFileKnownValue()).append(")"); //NON-NLS
+        }
+        query.append(" AND mime_type = '").append(mimeType).append("'");  //NON-NLS
+        return query.toString();
+    }
+
+    /**
      * Factory for populating the contents of the Media Sub Type Node with the
      * files that match MimeType which is represented by this position in the
      * tree.
@@ -404,21 +437,6 @@ public final class FileTypesByMimeType extends Observable implements AutopsyVisi
             super();
             addObserver(this);
             this.mimeType = mimeType;
-        }
-
-        /**
-         * Get children count without actually loading all nodes
-         *
-         * @return count(*) - the number of items that will be shown in this
-         *         items Directory Listing
-         */
-        private long calculateItems(SleuthkitCase sleuthkitCase, String mime_type) {
-            try {
-                return sleuthkitCase.countFilesWhere(createQuery(mime_type));
-            } catch (TskCoreException ex) {
-                LOGGER.log(Level.SEVERE, "Error getting file search view count", ex); //NON-NLS
-                return 0;
-            }
         }
 
         /**
@@ -442,34 +460,6 @@ public final class FileTypesByMimeType extends Observable implements AutopsyVisi
             return true;
         }
 
-        /**
-         * Create the portion of the query following WHERE for a query of the
-         * database for each file which matches the complete MIME type
-         * represented by this node. Matches against the mime_type column in
-         * tsk_files.
-         *
-         * @param mimeType - the complete mimetype of the file mediatype/subtype
-         *
-         * @return query.toString - portion of SQL query which will follow a
-         *         WHERE clause.
-         */
-        private String createQuery(String mimeType) {
-            StringBuilder query = new StringBuilder();
-            query.append("(dir_type = ").append(TskData.TSK_FS_NAME_TYPE_ENUM.REG.getValue()).append(")"); //NON-NLS
-            query.append(" AND (type IN (").append(TskData.TSK_DB_FILES_TYPE_ENUM.FS.ordinal()).append(",");  //NON-NLS
-            query.append(TskData.TSK_DB_FILES_TYPE_ENUM.CARVED.ordinal()).append(",");
-            query.append(TskData.TSK_DB_FILES_TYPE_ENUM.DERIVED.ordinal()).append(",");
-            if (!UserPreferences.hideSlackFilesInViewsTree()) {
-                query.append(TskData.TSK_DB_FILES_TYPE_ENUM.SLACK.ordinal()).append(",");
-            }
-            query.append(TskData.TSK_DB_FILES_TYPE_ENUM.LOCAL.ordinal()).append("))");
-            if (UserPreferences.hideKnownFilesInViewsTree()) {
-                query.append(" AND (known IS NULL OR known != ").append(TskData.FileKnown.KNOWN.getFileKnownValue()).append(")"); //NON-NLS
-            }
-            query.append(" AND mime_type = '").append(mimeType).append("'");  //NON-NLS
-            return query.toString();
-        }
-
         @Override
         public void update(Observable o, Object arg) {
             refresh(true);
@@ -485,43 +475,9 @@ public final class FileTypesByMimeType extends Observable implements AutopsyVisi
          */
         @Override
         protected Node createNodeForKey(Content key) {
-            return key.accept(new ContentVisitor.Default<AbstractNode>() {
-                @Override
-                public FileNode visit(File f) {
-                    return new FileNode(f, false);
-                }
-
-                @Override
-                public DirectoryNode visit(Directory d) {
-                    return new DirectoryNode(d);
-                }
-
-                @Override
-                public LayoutFileNode visit(LayoutFile lf) {
-                    return new LayoutFileNode(lf);
-                }
-
-                @Override
-                public LocalFileNode visit(DerivedFile df) {
-                    return new LocalFileNode(df);
-                }
-
-                @Override
-                public LocalFileNode visit(LocalFile lf) {
-                    return new LocalFileNode(lf);
-                }
-
-                @Override
-                public SlackFileNode visit(SlackFile sf) {
-                    return new SlackFileNode(sf, false);
-                }
-
-                @Override
-                protected AbstractNode defaultVisit(Content di) {
-                    throw new UnsupportedOperationException(NbBundle.getMessage(this.getClass(), "FileTypeChildren.exception.notSupported.msg", di.toString()));
-                }
-            });
+            return key.accept(new FileTypes.FileNodeCreationVisitor());
         }
+
     }
 
 }
